requirements.txt
openai==1.59.9
toml==0.10.2
python-dotenv==1.0.1
scikit-learn==1.6.1
spacy>=3.0,<4.0
nltk>=3.5,<4.0
pandas==2.2.3
beautifulsoup4==4.12.3

llama-index-vector-stores-deeplake==0.1.6
deeplake==3.9.18
llama-index==0.10.64
sentence-transformers==3.0.1

selenium
webdriver-manager